requirements.txt
django>=4.2.0
django-ninja>=1.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0
django-cors-headers
django-mptt